[!!!][TASK] Remove custom solution for something the core already does
The custom solution still threw PHP 8 related exceptions. 
This commit removes the custom solution in favour of the core solution to resolve file references.
This changes the type of the "processed(Assets|Images)" elements from an array to an object of type TYPO3\CMS\Core\Resource\FileReference.

This means that backend preview templates that accessed these files might have to be refactored.